OptionTank -> GenericTank; made GenericTank take an enum class extending IAttribute as a generic parameter
