+++ src/main/java/pro/trevor/tankgame/state/board/unit/tank/IAttribute.java
@@ -1,0 +1,5 @@
+package pro.trevor.tankgame.state.board.unit.tank;
+
+public interface IAttribute {
+    Class<?> getType();
+}
